+++ src/api/modules/crm/modules/client.service.ts
@@ -3,7 +3,7 @@ import { request } from '@/utils'
 /**销售管理-客户列表-新增客户**/
 export function httpBaseCrmClientCommonCreate(data: Omix) {
     return request({
-        url: `/api/windows/crm/client/common/create`,
+        url: `/api/consumers/create`,
         method: 'POST',
         data
     })
@@ -12,7 +12,7 @@ export function httpBaseCrmClientCommonCreate(data: Omix) {
 /**销售管理-客户列表-分页列表查询**/
 export function httpBaseCrmClientCommonConsumer(data: Omix) {
     return request({
-        url: `/api/windows/crm/client/common/consumer`,
+        url: `/api/consumers/column`,
         method: 'POST',
         data
     })